Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,9 +3,9 @@
 
 
 int main() {
-    int r;
+    float r;
     float pi=3.14;
-    scanf("%d\n",&r);
+    scanf("%f\n",&r);
     float area=pi*r*r;
     printf("Area: %.2f",area);
     return 0;
